feat(packages/core): Add File->Save as Notebook
This also reorders the New and Open File menu options to be more consistent with norms.

Fixes #5715
