+++ packages/core/src/main/save.ts
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2020 IBM Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { filters } from './open'
+import tellRendererToExecute from './tell'
+
+let previousChoice: string
+
+/**
+ * Save the current tab as a Notebook
+ *
+ */
+export default async function saveAsNotebook() {
+  const { dialog } = await import('electron')
+
+  const resp = await dialog.showSaveDialog({
+    title: 'Save as Notebook',
+    filters,
+    defaultPath: previousChoice
+  })
+
+  if (!resp.canceled) {
+    const thisChoice = resp.filePath
+    previousChoice = thisChoice
+    tellRendererToExecute(`snapshot "${thisChoice}"`, 'pexec')
+  }
+}
